[AR-3512] Enable configuration for rails requests (#23)
Logatron now allows for configuring additional fields to show in the
rails HTTP request log line. The configuration was made more simple
overall for use as well as adding a test for the railtie. The README was
updated with critical information about how logging works in rails as
well as explaining any magic that happens behind the scenes.