[AR-3547] Fix null 'id' field (#25)
The logatron log lines were printing the id field just fine when rails
was started via the command line. However, when ran in docker, the id
field was `null` each time logatron logged a line. After some
investigation, it was found that it had to do with the way unicorn and
logatron are loaded before rails when running in docker vs logatron being
loaded when within the rails context when starting on the command line.
To fix the problem, the logatron railtie must be included in the
`config/application.rb` after `Rails` is present but before bundler
requires all the things. The logatron railtie will insert its
'config.after_initialize` before the lograge railtie is loaded. This will
ensure the `config.after_initilize` block in the logatron railtie will
run before the one specified in lograge. The lograge
`config.after_initialize` block will then be guarenteed to have both
logatron and any lograge configuration complete before running.